+++ quam/components/channels.py
@@ -43,6 +43,7 @@ from qm.qua import (
     update_frequency,
     frame_rotation,
     frame_rotation_2pi,
+    time_tagging,
 )
 from qm.qua._dsl import (
     _PulseAmp,
@@ -201,6 +202,62 @@ class StickyChannelAddon(QuamComponent):
 
 
 @quam_dataclass
+class TimeTaggingAddon(QuamComponent):
+    """Addon to perform time tagging on a channel.
+
+    Args:
+        signal_threshold (float, optional): The signal threshold in volts.
+            If not specified, the default value is 800 / 4096 ≈ 0.195 V.
+        signal_polarity (Literal["above", "below"]): The polarity of the signal
+            threshold. Default is "below".
+        derivative_threshold (float, optional): The derivative threshold in volts/ns.
+            If not specified, the default value is 300 / 4096 ≈ 0.073 V/ns.
+        derivative_polarity (Literal["above", "below"]): The polarity of the derivative
+            threshold. Default is "below".
+
+    For details see [Time Tagging](https://docs.quantum-machines.co/latest/docs/Guides/features/#time-tagging)
+    """
+
+    signal_threshold: float = 800 / 4096
+    signal_polarity: Literal["above", "below"] = "below"
+    derivative_threshold: float = 300 / 4096
+    derivative_polarity: Literal["above", "below"] = "below"
+    enabled: bool = True
+
+    @property
+    def channel(self) -> Optional["Channel"]:
+        """If the parent is a channel, returns the parent, otherwise returns None."""
+        if isinstance(self.parent, Channel):
+            return self.parent
+        else:
+            return
+
+    @property
+    def config_settings(self):
+        if self.channel is not None:
+            return {"after": [self.channel]}
+
+    def apply_to_config(self, config: dict) -> None:
+        if self.channel is None:
+            return
+
+        if not self.enabled:
+            return
+
+        if self.signal_threshold is not None and abs(self.signal_threshold) > 1:
+            raise ValueError("TimeTaggingAddon.signal_threshold must be a voltage")
+        # TODO should we also check derivative threshold? What should the max value be?
+
+        ch_cfg = config["elements"][self.channel.name]
+        ch_cfg["outputPulseParameters"] = {
+            "signalThreshold": int(self.signal_threshold * 4096),
+            "signalPolarity": self.signal_polarity,
+            "derivativeThreshold": int(self.derivative_threshold * 4096),
+            "derivativePolarity": self.derivative_polarity,
+        }
+
+
+@quam_dataclass
 class Channel(QuamComponent, ABC):
     """Base QuAM component for a channel, can be output, input or both.
 
@@ -222,7 +279,6 @@ class Channel(QuamComponent, ABC):
 
     digital_outputs: Dict[str, DigitalOutputChannel] = field(default_factory=dict)
     sticky: Optional[StickyChannelAddon] = None
-
     intermediate_frequency: Optional[float] = None
     thread: Optional[str] = None
 
@@ -631,6 +687,8 @@ class InSingleChannel(Channel):
     time_of_flight: int = 24
     smearing: int = 0
 
+    time_tagging: Optional[TimeTaggingAddon] = None
+
     def apply_to_config(self, config: dict):
         """Adds this InSingleChannel to the QUA configuration.
 
@@ -879,6 +937,65 @@ class InSingleChannel(Channel):
             ),
         )
         return tuple(qua_vars)
+
+    def measure_time_tagging(
+        self,
+        pulse_name: str,
+        size: int,
+        max_time: QuaNumberType,
+        qua_vars: Optional[Tuple[QuaVariableType, QuaNumberType]] = None,
+        stream: Optional[StreamType] = None,
+        mode: Literal["analog", "high_res", "digital"] = "analog",
+    ) -> Tuple[QuaVariableType, QuaNumberType]:
+        """Perform a time tagging measurement on this channel.
+
+        For details see https://docs.quantum-machines.co/latest/docs/Guides/features/#time-tagging
+
+        Args:
+            pulse_name (str): The name of the pulse to play. Should be registered in
+                `self.operations`.
+            size (int): The size of the QUA array to store the times of the detected
+                pulses. Ignored if `qua_vars` is provided.
+            max_time (QuaNumberType): The maximum time to search for pulses.
+            qua_vars (Tuple[QuaVariableType, QuaNumberType], optional): QUA variables
+                to store the times and counts of the detected pulses. If not provided,
+                new variables will be declared and returned.
+            stream (Optional[StreamType]): The stream to save the measurement result to.
+                If not provided, the raw ADC signal will not be streamed.
+            mode (Literal["analog", "high_res", "digital"]): The time tagging mode.
+
+        Returns:
+            times (QuaVariableType): The QUA variable to store the times of the detected
+                pulses.
+            counts (QuaNumberType): The number of detected pulses.
+
+        Example:
+            ```python
+            times, counts = channel.measure_time_tagging("readout", size=1000, max_time=1000)
+            ```
+        """
+        if mode == "analog":
+            time_tagging_func = time_tagging.analog
+        elif mode == "high_res":
+            time_tagging_func = time_tagging.high_res
+        elif mode == "digital":
+            time_tagging_func = time_tagging.digital
+        else:
+            raise ValueError(f"Invalid time tagging mode: {mode}")
+
+        if qua_vars is None:
+            times = declare(int, size=size)
+            counts = declare(int)
+        else:
+            times, counts = qua_vars
+
+        measure(
+            pulse_name,
+            self.name,
+            stream,
+            time_tagging_func(target=times, max_time=max_time, targetLen=counts),
+        )
+        return times, counts
 
 
 @quam_dataclass